Add simple job-running code
Add a simple prototype that can execute a subprocess in the
sandbox.  This requires passing LD_PRELOAD in the environment to
ensure that the sandbox is loaded as well as setting up the
shared memory region that passes parameters down to the sandbox.

A test application called "runner" has been added to demonstrate
the sandbox in action.