Implement interception of open() in the sandbox
Make the sandbox intercept calls to open(2) and pass them back to
factory to be vetted.  The open(2) syscall will only be made if
factory allows the request.  As a proof of concept, factory
currently rejects everything.

This has required writing the following:

- The EventLoop is a C++ wrapper around libevent that takes events
and dispatches them to the proper handler
- The JobManager is responsible for launching new processes and
tracing them as a Job.  While we only ever launch a single process
as a job, note that the child process could launch multiple
processes itself and therefore there will could be multiple
processes for a single job.  A Job is the unit of granularity for
enforcement of file access.
- A MsgSocket is used for two-way communication between a sandbox
and factory.  Because we don't know what processes a job has
spawned, we require the sandbox to explicity inform us of its
job ID before doing anything else.  MsgSockets are owned by
jobs.  In between the acceptance of a new socket and it being
mapped to a job, the MsgSocketManager owns the socket while it
is in this "incomplete" state.