Add dependency tracking between jobs
Add a framework for tracking outstanding Products to be built.
Products represent files used in the build.  PendingJobs represent
the process that must be run to consume input Products to produce
one or more output Products.  When a Product is built, it notifies
the Products that consume it that it is complete.  When all
Products consumed by a Product are complete, its PendingJob is
enqueued to be run by the JobManager.