Handle the case where a granddepenency is modified.
If A depends on B and B depends on C, and A is newer than B but
B is older than C, then B should be rebuilt, triggering a rebuild
of A.  Unfortunately we were only checking that A was newer than B
at the start and not queuing it to be rebuilt after B was rebuilt.

Fix this by forwarding the NeedsBuild property to all dependees
of a product when we determine it must be built.