Properly handle directory products
Directory products need to be handled specially in two ways:

- Permissions need to be granted to read/write directory trees
  rather than the exact path
- If a product depends on a directory, it should not be rebuilt
  if the directory is newer than the product.  The reason for
  this is that products must depend on their parent directory
  (to ensure that their parent directory is exists before we
  try to create the file), but the parent dir's mtime changes
  every time that any file in the dir is written to

Fix this by requiring the lua script to tell us if any product
is a directory.

This feels too heavyweight and probably needs to be re-thought.
In theory, the lua script only needs to tell us if a product we
will create is a directory (we can check if an existing path
represents a directory).  But fixing that would require a re-work
of how products are managed, as there is no guarantee about the
order between when a product's command is defined (which is when
we would discover that the product is a directory) compared to
when a different product declares a dependency on that directory.