Fix build steps being missed
Because NeedsBuild() is transitive, it's not easy to build up
a list of products that need a build as we iterate through the
list.  We were only putting the top-level product that needed
a build in the list of products that had outstanding actions.

Rather than trying to fix this, just iterate over the list of
all products a second time to find the products that are
outstanding.