Don't leak the tmpdir if there is no work
Calling exit() rather than returning here means that the
TempFileManager's destructor never runs and removes the tempdir.
Fix this one easy case.  There are many, many more difficult cases
to handle in the error paths in the interpeter (where we cannot
throw an exception because it can't propagate through the lua
C code).