Allow multiple config files to be processed together
Change the semantics of factory.include_config: if it is passed
multiple config files in a list, all of their configs will be
collected together and passed to a single invocation of a
lua callback.  Definitions are now allow to specify a list
of handled ingest types.  If they specify more than one, they
will only be invoked on a include_config() call that creates
config matching all of the types.  The types must be defines
by the files in the same order as in the definition.

For example, if an ingest that processes ["foo", "bar"] is defined,
it will only be called after a call to:

factory.include_config(parent_config, fooFile, barFile)

(assuming that fooFile defines a "foo" config and barFile defines
a "bar" config)

There is currently a restriction that when multi-including config,
each config file may define only a single top-level key.