Add declarative method for parsing lua table
Add a method that allows the programmer to simply declare the
expected format of a complicated lua data structure that they
expect to receive.  The new infrastructure will validate that
the data structure passed from the lua code meets the expected
format (including validating that there are no unexpected fields).

This should greatly simplify the process of accepting structured
data from lua.