Implement support for rewriting a blocking wait4(2)
Implement a blocking wait4(2) by using the new deferred program
support.  A deferred pdwait4() call is made that exits from the
first program, leaves the context of the EBPF VM and then performs
the blocking call.  When the call completes the EBPF code re-enters
the VM and calls the deferred program to return results to
userland.