Add hack for directory write access
Directories can only be opened with O_RDONLY.  If we try to open
a file and get EISDIR, re-open it with O_RDONLY to make the
kernel happy.  The capabilities that we set on the descriptor
will still allow files opened with openat() with the directory
descriptor to have write access.