Don't preopen FD for writing
Preopening a file descriptor for writing is tricky because there
are a bunch of flags that you have to emulate properly (e.g.
O_TRUNC, O_CREAT, etc).  Instead preopen its parent directory
and change the syscall rewriter to validate that the file that
we are trying to open is the one that we have permission for.

This opens up a hole where a malicious application inside of the
sandbox could use the directory file descriptor to gain access
to files that they shouldn't.  This can be fixed in the future
by adding an fphashtable map type to EBPF that holds file pointers
directory, allowing us to run the sandboxed application without
the directory open in its FD table.