Allow multiple files in a dir to be opened
The change to preopen parent directories rather than files
introduced a bug where we could not grant permission to access
multiple files in a directory (other than granting permission to
the whole directory).  Similarly, we couldn't grant permission to
a directory and a file inside of it, which could arise if you
wanted different permissions on the dir and the file.

Fix this by having the file_lookup_map contain the full path name.
This allows us to do lookups on different files in the same dir
and get different FDs with different capabiliies.