Do manual symlink evaluation in syscall rewriting
When linking, it's common to come across symlinks that go from
one directory to which we have permission to access to another
(e.g. /usr/lib/libcxxrt.so is a symlink to lib in /lib).
In capability mode, we are not allowed to resolve such symlinks
in the standard syscall implementations.

Fix this by manually resolving symlinks on file operations
in the syscall rewriters.  At each step, so long as the target
of the symlink falls within a directory we have access to, we
can continue the operation.

Currently we handle up to 3 lookups (symlink to symlink to real
file), as that's the worst case I've seen in linking so far
(libpthread.so -> libthr.so -> libthr.so.3).