Canonicalize paths passed to syscalls
ebpf_map_lookup_path() can only match a path if it is canonical
form (assuming that the map was initialized with paths in
canonical form).  Canonicalize paths passed to syscalls before
trying to look them up to ensure that we match on non-canonical
paths to files that the sandboxed application has the right to
access.

Re-use the scratch buffer used for canonicalization in symlink
resolution so that we don't waste a scratch buffer.