Allow opening of relative paths when cwd is not pre-opened
In some cases, we may want to allow the sandboxed process to
access some files in their work dir but not everything beneath
the work_dir.  Cache the path name of their cwd in a new map
and use that to generate an absolute path when we do a relative
path lookup.

Lookup via an fd pointing to the cwd is still supported, as any
fchdir operation would mean that we no longer know what the path
to cwd is.