Fix some usages of exec
Sometimes exec will be called with a full path to the binary in
fname but just its basename in argv[0].  When this happens, rtld
is not able to open the program and we fail to run it.

Work around it by replacing argv[0] with a full path to the
binary.