Prefer relative path lookups by path rather than cwd fd
It is entirely possible to have fewer permissions on your cwd
than a specific file that is a descendant of the cwd.  Handle this
by prefering to do a lookup via path rather than the fd for cwd.
Fall back to using cwd for cases where we have done an fchdir.