syscall rewriters: optimize for no symlink case
Rather than doing an expensive manual attempt to resolve symlinks
in case they point outside of the current directory FD,
speculatively try to open with the first dir FD that we find and
fall back on symlink resolution if that fails with ENOTCAP.  This
should improve performance in the common case of no symlinks
or symlinks that only point within the current dir FD.

Cases where symlinks point outside of the current dir FD, or cases
where we got ENOTCAP for reasons other than a symlink (e.g. trying
to open for write if we only have a read capability), will see
a mild pessimization.

As a side effect, this fixes a bug in our handling of
AT_SYMLINK_NOFOLLOW.  This flag means "don't resolve a symlink if
it is the last component of a path" but we we interpreting it to
mean "don't resolve any symlinks at all".