Detect existing directories
Some directories may not have any command products in them
(for instance, directories in the source tree).  In such cases
the directories must exist before factory is run, so check for a
directory when we create the product.

Ensure that we do not create a dependency between a product and
a directory.