Handle corner case in access(2) rewriter
It is possible for access("foo", X) to be called on a preopened
directory.  In this case, we can't emulate the system call for
all values of X without additional kernel support, but it is
easy to implement 0 (which is a simple test of existence).  As
we already have the directory opened, the directory must exist.