Working support for conditional compilation
Add support for conditionally setting build configuration.
Any config nodes that contain a list are not split into
two parts: the "std" node, which contains a list of all values
that are always included in the build, and an "opt" node, which
contains a list of optional values.  Options have both a match
criteria that defines when the config is active as well as the
value itself.